+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Note that version stamps correspond to Git tags. There's no version file.
 
+**Version v1.1.0**
+
+- Added a couple cells that demonstrate that `TypedDataset` performance
+  is generally equivalent to native Spark DataFrame performance.
+
 **Version v1.0.0**
 
 - As delivered at Scala.IO on 31 October, 2018.